requirements.txt
pymupdf==1.18.19
pytesseract
pillow
spacy
